Add quarter-star rating support
API:
- Changed UserRating from int? to double? in Item, UpdateItemRequest,
  and TopItem models to support fractional ratings (0.25 increments)

Frontend:
- New StarRating component with SVG clip-path rendering for partial
  star fills (quarter, half, three-quarter)
- Interactive mode: tap position within a star determines quarter
  increment (e.g., tap left edge = 0.25, middle = 0.5, right = 1.0)
- Three sizes: sm (lists), md (detail view), lg (edit mode)
- Replaced all hardcoded star text/SVGs across ItemDetailView,
  ItemsView, and StatsView with the StarRating component
- Numeric rating displayed alongside stars for precision

+++ src/api/Models/ApiModels.cs
@@ -74,7 +74,7 @@ public class UpdateItemRequest
     public VenueInfo? Venue { get; set; }
 
     [JsonPropertyName("userRating")]
-    public int? UserRating { get; set; }
+    public double? UserRating { get; set; }
 
     [JsonPropertyName("userNotes")]
     public string? UserNotes { get; set; }
